bigquery_medallion_migration/functions/create_ml_snapshot/requirements.txt
functions-framework==3.5.0
google-cloud-bigquery==3.14.1
pandas==2.1.4
